refactor : 검증로직 CalculatorValidator로 분할 및 Operator enum 클래스 내부에서 파일로 분할 - Calculator   1. 불필요한 상수 (FIRST_INDEX) 제거 및 인스턴스 변수 네이밍 수정 (LAST_INDEX -> lastIndex)   2. 문자열 전처리 과정 위치 변경 (validateUserInput 내부에 split와 lastIndex를 가져오는 로직이 있었는데 enterFormula() 내부로 위치변경)      - 변경 이유 : split하여 list를 생성하고 lastIndex를 가져오는 부분은 검증 로직 내부에서 처리하는 것이 아니라 문자열을 입력받는 함수에서                    처리하는게 맞는 책임일거 같다.   3. validateFormula() 함수를 생성하여 검증하는 부분의 책임을 CalculatorValidator로 넘겨줌. - CalculatorValidator   1. isValidate...와 같은 명확하지 않은 메소드 네이밍 수정 - Operator   1. 기존 Calculator 클래스 내부에 있던 enum 분할 -> 계산 부분을 이 enum에게 책임을 넘길 예정 (TODO)
