+++ src/main/java/Main.java
@@ -3,7 +3,7 @@ public class Main {
 
   public static void main(String[] args) {
       Calculator calculator = new Calculator();
-      calculator.enterFormula("1 / 0");
-      calculator.calculate();
+      calculator.enterFormula("2 * 3 + 5");
+      System.out.println(calculator.calculate());
   }
 }
