+++ src/main/java/Operator.java
@@ -1,0 +1,12 @@
+public enum Operator {
+  PLUS("+"),
+  MINUS("-"),
+  MULTIPLE("*"),
+  DIVIDE("/");
+
+  public final String symbol;
+
+  Operator(String symbol) {
+    this.symbol = symbol;
+  }
+}
